src/requirements.txt
numpy==1.23.0
Pillow==10.0.0
pyotp==2.9.0
qrcode==7.3.1
qreader==2.13
typer==0.9.0
